requirements.txt
aerich==0.6.3
aiosqlite==0.17.0
anyio==3.6.1
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
dictdiffer==0.9.0
ecdsa==0.18.0
fastapi==0.79.0
h11==0.13.0
httptools==0.4.0
idna==3.3
iso8601==1.0.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
pypika-tortoise==0.1.6
python-dotenv==0.20.0
python-jose==3.3.0
pytz==2022.1
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tomlkit==0.11.2
tortoise-orm==0.19.2
typing_extensions==4.3.0
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
